+++ lib/plugins/aws/package/compile/events/cognitoUserPool.js
@@ -2,7 +2,7 @@
 
 const _ = require('lodash');
 const BbPromise = require('bluebird');
-const { addCustomResourceToService } = require('../../../../customResources');
+const { addCustomResourceToService } = require('../../../customResources');
 
 const validTriggerSources = [
   'PreSignUp',
