+++ lib/plugins/index.js
@@ -47,7 +47,7 @@ module.exports = [
   require('./aws/package/compile/events/iot/index.js'),
   require('./aws/package/compile/events/cloudWatchEvent.js'),
   require('./aws/package/compile/events/cloudWatchLog.js'),
-  require('./aws/package/compile/events/cognitoUserPool/index.js'),
+  require('./aws/package/compile/events/cognitoUserPool.js'),
   require('./aws/package/compile/events/eventBridge/index.js'),
   require('./aws/package/compile/events/sqs/index.js'),
   require('./aws/package/compile/events/cloudFront.js'),
